requirements.txt
wheel
Flask
Redis
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-pubsub
requests